Remove libakonadi-qt4
There is already akonadi for this.

The only package that was depending on this one was kdepimlibs4, which had
already been updated in Git back in May not to depend on it, but for some
reason the version in the repositories was still the previous version. So
I just built and uploaded the Git version, so that no package depends on
libakonadi-qt4 anymore.